+++ src/main/java/com/enn3developer/gtnhvoice/ClientProxy.java
@@ -5,6 +5,7 @@ import com.enn3developer.gtnhvoice.client.ClientConnectionEventHandler;
 import com.enn3developer.gtnhvoice.client.VoiceClientManager;
 import com.enn3developer.gtnhvoice.client.VoiceListenerTickHandler;
 import com.enn3developer.gtnhvoice.client.capture.CaptureManager;
+import com.enn3developer.gtnhvoice.client.hud.VoiceHudRenderer;
 
 import cpw.mods.fml.common.event.FMLPreInitializationEvent;
 
@@ -14,6 +15,7 @@ public class ClientProxy extends CommonProxy {
     private final ActivationGate activationGate = new ActivationGate();
     private final ClientConnectionEventHandler clientConnectionEventHandler = new ClientConnectionEventHandler();
     private final VoiceListenerTickHandler voiceListenerTickHandler = new VoiceListenerTickHandler();
+    private final VoiceHudRenderer voiceHudRenderer = new VoiceHudRenderer();
 
     @Override
     public void preInit(FMLPreInitializationEvent event) {
@@ -22,6 +24,7 @@ public class ClientProxy extends CommonProxy {
         activationGate.register();
         clientConnectionEventHandler.register();
         voiceListenerTickHandler.register();
+        voiceHudRenderer.register();
 
         VoiceClientManager.getInstance()
             .bindCaptureManager(captureManager);